Fix show comments button
The "show comments" is now fixed, it is created only once and display
the comments correctly, everything without the help of an interval
timer anymore.

When clicking on "show comments", as the button was recreated at every
interval, the comments were constantly disappearing and the button not
fading away.

The code was restructured into sub-function. This is temporary. Then the
videos concerning function got the comments-side operations triggered
only if the comments are loaded (detected via the number of tag that
have a specific class when they do loading thing ; the comments section
one is deleted when the comments finished loading) via an interval.

The code is pretty, the next commit will correct this.