Ongoing script restruct
Add ./impartial-youtube-restruct.user.js file, which is the same as the
./impartial-youtube.user.js but being rewritten in a clearer way.

The ./impartial-youtube.user.js script partially works but it has some
useless subfunc and other dirty artifacts.

The new version is based on one big if/else which triggers it's content
according to the url.
As it is not tested and polished, this restructuration is seperate from
the main file, and no branching is done due to the fact that I'm hardly
referring myself to the original script.