Modify show comments button
Modify show comments button to make it appears at the loading of the
page instead of an awkward interval check to create it.

Doing it this way is more stable, the only downside is that the user
won't be able to comment the video without pushing the show comments
button.